+++ packages/server-wallet/src/__test-with-peers__/utils.ts
@@ -1,7 +1,9 @@
-import {ChannelResult} from '@statechannels/client-api-schema';
+import {ChannelResult, CreateChannelParams} from '@statechannels/client-api-schema';
 import _ from 'lodash';
 
 import {Engine} from '..';
+import {participantA, participantB} from '../../jest/with-peers-setup-teardown';
+import {createChannelArgs} from '../engine/__test__/fixtures/create-channel';
 
 export async function expectLatestStateToMatch(
   channelId: string,
@@ -10,4 +12,11 @@ export async function expectLatestStateToMatch(
 ): Promise<void> {
   const latest = await engine.getState({channelId});
   expect(latest.channelResult).toMatchObject(partial);
+}
+
+export function getWithPeersCreateChannelsArgs(): CreateChannelParams {
+  return createChannelArgs({
+    participants: [participantA, participantB],
+    fundingStrategy: 'Fake',
+  });
 }
